packages/clerk/pyproject.toml
[project]
name = "openclerk"
version = "0.1.0"
description = "Community Library of Executable Reasoning Kits"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
authors = [
    {name = "OpenClerk Contributors"}
]
keywords = ["llm", "ai", "workflow", "reasoning", "langchain", "langgraph"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "langchain>=1.2.4",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.6",
    "python-dotenv>=1.2.1",
    "sqlalchemy[asyncio]>=2.0.36",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "supabase>=2.11.0",
    "pypdf>=5.1.0",
    "openpyxl>=3.1.5",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "itsdangerous>=2.2.0",
    "sse-starlette>=2.1.0",
    "pgvector>=0.3.6",
    "langchain-postgres>=0.0.17",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "mcp>=1.26.0",
    "langchain-anthropic>=1.3.5",
    "langchain-mistralai>=1.1.2",
    "langchain-google-genai>=4.2.1",
    "langchain-google-vertexai>=3.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
    "mypy>=1.10.0",
]

[project.scripts]
clerk = "openclerk.cli:main"
openclerk = "openclerk.cli:main"

[project.urls]
Homepage = "https://openclerk.dev"
Documentation = "https://openclerk.dev/docs"
Repository = "https://github.com/yourorg/openclerk"
Issues = "https://github.com/yourorg/openclerk/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/openclerk"]
exclude = [
    "**/tests/",
    "**/test_*.py",
]

[tool.uv]
package = true
